Add Remark Command Functionality
Remark Command does not function as the back end code is not implemented. Users are not able to add remarks as the system only returns a message.

Add Remark Command functionality and test cases. Remark command updates and refreshes UI after the command is run, adding a remark to the user indicated.